homework done

+++ 2024-01/2024-01-30-H/Point.cpp
@@ -1,6 +1,6 @@
 #include "Point.h"
 
-Point::Point(int x, int y)
+Point::Point(float x, float y)
 	:x(x), y(y)
 {
 }
